test cases changed accordingly
